Bug #1339534 pwm: sprd: duty_ns * PWM_MOD_MAX overflow
[root cause  ] duty_ns * PWM_MOD_MAX overfllow
[changes     ] turn into duty_ns u64
[side effects]no
[self test   ]:no
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:yes
[confirm dependent commit]:no
[board]:common
[test case]:boot
[reviewers   ] chris.wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

duty_ns * PWM_MOD_MAX overflow,change duty_ns type.

+++ drivers/pwm/pwm-sprd.c
@@ -74,7 +74,7 @@ static int sprd_pwm_config(struct pwm_chip *chip, struct pwm_device *pwm,
 		spc->eb_enabled[pwm->hwpwm] = true;
 	}
 
-	tmp = duty_ns * PWM_MOD_MAX;
+	tmp = (u64)duty_ns * PWM_MOD_MAX;
 	level = DIV_ROUND_CLOSEST_ULL(tmp, period_ns);
 	dev_dbg(chip->dev, "duty_ns = %d, period_ns = %d, level = %d\n",
 		duty_ns, period_ns, level);
